refactor(tests): fold micro-files into per-module homes
73 -> 52 files in tests/unit with identical coverage: the twelve
per-model test_yaml_model_* files merge into test_yaml_model_equivalence
(plainconvunet and segresnet stay standalone -- size cap and a
same-named test), eleven micro-files fold into the per-module home of
the code they test, and package/import/wheel checks merge into one
test_packaging.py. Moves are verbatim (AST-identical against HEAD,
private helpers renamed only on collision); 1087 tests collected before
and after, full suite green.
